LPD-31094 portal-search-elasticsearch7-impl: Move tests to the right class because it is testing query translation
